Parte Admin terminada Parte Aux terminada

+++ Angular/src/app/components/admin/dashboard/asigAuxiliar/asignacionAuxiliar.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { AsignacionAuxiliarService } from '../../../../services/asignacionAuxiliar.service';
+import { FormBuilder, FormGroup, Validators, FormControl } from '@angular/forms';
 
 @Component({
   selector: 'app-asignacionauxiliar',
@@ -8,12 +9,16 @@ import { AsignacionAuxiliarService } from '../../../../services/asignacionAuxili
 export class AsignacionAuxiliar implements OnInit {
   arrayAsignacion: any[] = [];
   varAuxiliar: any;
-  constructor(private service: AsignacionAuxiliarService) { }
+  loginForm: FormGroup;
+  submitted = false;
+  constructor(private service: AsignacionAuxiliarService, private formBuilder: FormBuilder) { }
 
   ngOnInit() {
     this.inicializar();
+    this.loginForm = this.formBuilder.group({
+      motivo: ['', Validators.required]});
   }
-
+  get f() { return this.loginForm.controls; }
   inicializar() {
     this.service.get().subscribe(data => {
       this.arrayAsignacion = data;
@@ -21,6 +26,11 @@ export class AsignacionAuxiliar implements OnInit {
   }
 
   borrar(id: any) {
+    this.submitted = true;
+    // stop here if form is invalid
+    if (this.loginForm.invalid) {
+      return;
+    }
     this.service.delete(id)
     .subscribe(res => {
       this.inicializar();
